% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCoop.R
\name{bayesCoop}
\alias{bayesCoop}
\title{bayesCoop}
\usage{
bayesCoop(
  data_train,
  data_test,
  family = "gaussian",
  ss = c(0.05, 1),
  group = TRUE,
  bb = TRUE,
  alpha_dirich = 1,
  bbiters = 1100,
  bbburn = 100,
  maxit = 100,
  filter = TRUE,
  abd_thresh = 0,
  prev_thresh = 0.1,
  Warning = TRUE,
  verbose = TRUE,
  control = list()
)
}
\arguments{
\item{data_train}{a list of feature_table, sample_metadata and feature_metadata from training data(unstandardized). See \emph{IntegratedLearner} for more details.}

\item{data_test}{a list of feature_table, sample_metadata and feature_metadata from testing data(unstandardized). See \emph{IntegratedLearner} for more details..}

\item{family}{currently supports only Gaussian family. Default value: "gaussian".}

\item{ss}{a two-dimensional vector representing the scale parameters of the spike and slab parts, namely, s0 and s1 respectively, with s0 < s1. Default value: ss = c(s0, s1) = c(0.05, 1).}

\item{group}{Logical. If \code{TRUE}, \code{group} is a list of variable names where \code{group[[k]]} includes variables in the \eqn{k}-th group. For ungrouped predictors, the prior is double-exponential with scale \code{ss[2]} and mean 0. If \code{FALSE}, \code{group = NULL}. Default is \code{TRUE}.}

\item{bb}{logical. If TRUE, Bayesian bootstrap is implemented, else, maximum a posteriori (MAP) estimation is implemented which requires additional parameters to be passed to \code{control}. Default value: TRUE.}

\item{alpha_dirich}{shape parameter for Dirichlet distribution. Default value: 1.}

\item{bbiters}{number of Bayesian bootstrap iterations. Default value: 1100.}

\item{bbburn}{number of burn-in samples. Default value: 100.}

\item{maxit}{integer giving the maximal number of EM iterations. Default value: 100.}

\item{filter}{logical. If TRUE, features are filtered as per abundance and prevalence thresholds. Default value: TRUE.}

\item{abd_thresh}{threshold for abundance of features. Default value: 0.}

\item{prev_thresh}{threshold for prevalence of features. Default value: 0.1.}

\item{Warning}{logical. If TRUE, show the error messages of not convergence. Default value: TRUE.}

\item{verbose}{logical. If TRUE, print out number of iterations and computational time. Default value: TRUE.}

\item{control}{a named list with one element \code{rho}, which is a numeric vector of user-specified rho values.}
}
\value{
a list containing the following components is returned if \code{bb} is TRUE.
\item{y_pred}{predicted values of response}
\item{mspe}{mean square prediction error}
\item{beta_samples}{approximate posterior samples of beta}
\item{beta_postmed}{posterior median as an estimate for beta}
\item{rho_postmed}{posterior median as an estimate for rho}
\item{time}{time taken for BayesCoop to run}
a list containing the following components if \code{bb} is FALSE.
\item{y_pred}{predicted values of response}
\item{mspe}{mean square prediction error}
\item{beta_MAP}{MAP estimate of beta}
\item{rho_MAP}{optimal rho among the specified values chosen corresponding to minimum mean square prediction error, returns the same value if only one value for rho is specified in \code{control}}
\item{time}{time taken for MAP estimation to run}
}
\description{
This function implements the BayesCOOP methodology for supervised multimodal integration. It combines jittered group spike-and-slab LASSO regularization with intermediate fusion to enable integrative learning across multiple data modalities. For uncertainty quantification, BayesCOOP applies the Bayesian bootstrap to generate approximate posterior samples by performing maximum a posteriori (MAP) estimation on jittered and resampled multimodal datasets. Currently, only continuous outcomes are supported.
}
