% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BayesCOOP.R
\name{predict.BayesCOOP}
\alias{predict.BayesCOOP}
\title{predict.BayesCOOP}
\usage{
\method{predict}{BayesCOOP}(
  object,
  newdata,
  family = "gaussian",
  bb = TRUE,
  warning = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{fitted object returned by the function \code{\link{BayesCOOP}} implemented on the training dataset.}

\item{newdata}{a list of feature_table, sample_metadata and feature_metadata from validation dataset (unstandardized). See \emph{IntegratedLearner} for more details.}

\item{family}{currently supports only Gaussian family. Default: "gaussian".}

\item{bb}{logical. If TRUE, posterior predictive sampling is performed using the posterior median of \eqn{rho} stored in the fitted object; if FALSE, point prediction is provided for MAP estimate of \eqn{rho} stored in the fitted object. Default: TRUE.}

\item{warning}{logical. If TRUE, emit non-convergence warnings. Default: TRUE.}

\item{verbose}{logical. If TRUE, print iteration counts and runtime. Default: TRUE.}

\item{...}{further arguments passed from \code{predict()}.}
}
\value{
If \code{bb = TRUE}, a list with:
\item{y_samples}{posterior predictive draws (approximate)}
\item{y_pred}{posterior predictive median for each held-out sample}
\item{y_valid}{True values of response in the validation dataset}

If \code{bb = FALSE}, a list with:
\item{y_pred}{predicted response values for data}
\item{y_valid}{True values of response in the validation dataset}
}
\description{
This function predicts the response from a fitted object using BayesCOOP methodology. Currently, only continuous outcomes are supported.
}
