% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCOOP.R
\name{BayesCOOP}
\alias{BayesCOOP}
\title{Scalable Bayesian Cooperative Learning for Multimodal Integration}
\usage{
bayesCoop(data_train, data_test, family = "gaussian", 
                     ss = c(0.05, 1), group = TRUE,
                     alpha_dirich = 1, maxit = 100, 
                     bbiters = 1100, bbburn = 100,
                     abd_thresh = 0, prev_thresh = 0.1,
                     Warning = TRUE, verbose = TRUE)
}
\arguments{
\item{data_train}{A list of feature_table, sample_metadata and feature_metadata from training data..}
\item{data_test}{A list of feature_table, sample_metadata and feature_metadata from testing data.}
\item{family}{Currently supports only Gaussian family. Default value: "gaussian".}
\item{ss}{(s0, s1): s0 and s1 respectively stands for the scale parameters of the spike and slab parts, s0 < s1. Default value: s0 = 0.05, s1 = 1.}
\item{group}{logical. If TRUE, group is a list of variable names, group[[k]] includes variables in the k-th group. For ungrouped predictors, the prior is double-exponential with scale ss[2] and mean 0. If FALSE, group = NULL. Default value: TRUE.}
\item{alpha_dirich}{Shape parameter for Dirichlet distribution. Default value: 1.}
\item{maxit}{Integer giving the maximal number of EM iterations.}
\item{bbiters}{Number of Bayesian bootstrap iterations. Default value: 1100.}
\item{bbburn}{Number of burn-in samples. Default value: 100.}
\item{abd_thresh}{Threshold for abundance of features. Default value: 0.}
\item{prev_thresh}{threshold for prevalence of features. Default value: 0.1.}
\item{Warning}{Default is TRUE.}
\item{verbose}{Default is TRUE.}
}
\value{
A list containing approximate pseudo-posterior samples of the regression coefficients and hyperparameters:
\itemize{
\item beta_postmed: vector of size px1, posterior median estimate of the beta.
\item rho_postmed: scalar, posterior median estimate of the agreement penalty $\rho$.
\item beta_samples: matrix of size npost x p where npost is the number of posterior samples. Returns  approximate posterior samples of beta.
\item y_pred: a vector. Returns the predicted value of y.
\item mspe: a scalar. Returns the mean squared prediction error.
\item time: a scalar. Returns the time needed to run the procedure. 
}
}
\description{
\code{BayesCOOP} provides the analysis codes from the Bayesian multiview project.
}
\details{
This function implements the BayesCOOP methodology for supervised multimodal integration. 
It combines jittered group spike-and-slab LASSO regularization with intermediate fusion to 
enable integrative learning across multiple data modalities. For uncertainty quantification,
BayesCOOP applies the Bayesian bootstrap to generate approximate posterior samples by 
performing maximum a posteriori (MAP) estimation on jittered and resampled multimodal datasets. 
Currently, only continuous outcomes are supported.
}
\examples{
rm(list = ls())

########################
###### libraries #######
########################

#library(remotes)
#install_github("nyiuab/BhGLM", force=T, build_vignettes=T)
library(BhGLM)
library(tidyverse)
library(caret)

#########################################
###### sourcing required functions ###### 
#########################################

source("https://github.com/himelmallick/BayesCOOP/tree/master/R/bayesCoop.R")

#########################################################################################
######################### Loading the StelzerDOS real dataset ###########################
#########################################################################################

data_train = get(load(url("https://raw.githubusercontent.com/himelmallick/IntegratedLearner/master/data/StelzerDOS.RData"))); rm(pcl)
data_test = get(load(url("https://raw.githubusercontent.com/himelmallick/IntegratedLearner/master/data/StelzerDOS_valid.RData"))); rm(pcl)

#########################################################################################
##### Pre-processing the longitudinal data by considering only baseline observations ####
#########################################################################################

# Remove metabolomics from the train set to match with validation
data_train$feature_metadata = data_train$feature_metadata %>% dplyr::filter(featureType!='Metabolomics')
data_train$feature_table = data_train$feature_table[rownames(data_train$feature_metadata),]

# Consider only baseline observations for the train set
positions = grep("A", colnames(data_train$feature_table), ignore.case = TRUE)
data_train$feature_table = data_train$feature_table[, positions]
data_train$sample_metadata = data_train$sample_metadata[positions, ]
rm(positions)

# Consider only baseline observations for the validation set
positions = grep("G1", colnames(data_test$feature_table))
data_test$feature_table = data_test$feature_table[, positions]
data_test$sample_metadata = data_test$sample_metadata[positions, ]
rm(positions)

#########################################################################################
############################## BayesCoop Implementation #################################
#########################################################################################

set.seed(1)
result = bayesCoop(data_train, data_test, family = "gaussian", 
                   ss = c(0.05, 1), group = TRUE,
                   alpha_dirich = 1, maxit = 100, 
                   bbiters = 1100, bbburn = 100,
                   abd_thresh = 0, prev_thresh = 0.1,
                   Warning = TRUE, verbose = TRUE)

result
}
